Removed Bounteous references in favor of Adobe. Updated README and added other doc files (LICENSE, CONTRIBUTING, CODE_OF_CONDUCT)
