+++ aem-component-generator/.github/CONTRIBUTING.md
@@ -1,0 +1,24 @@
+# Contributing
+
+Thanks for choosing to contribute!
+
+The following are a set of guidelines to follow when contributing to this project.
+
+## Code Of Conduct
+
+This project adheres to the Adobe [code of conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold
+this code. Please report unacceptable behavior to the team.
+
+## Contributor License Agreement
+
+All third-party contributions to this project must be accompanied by a signed contributor license agreement. This gives
+Adobe permission to redistribute your contributions as part of the project.
+[Sign our CLA](http://opensource.adobe.com/cla.html). You only need to submit an Adobe CLA one time, so if you have
+submitted one previously, you are good to go!
+
+## How to contribute
+
+New code contributions should be made primarily using GitHub pull requests. This involves creating a fork of the
+project in your personal space, adding your new code in a branch and triggering a pull request.
+
+See how to perform pull requests at https://help.github.com/articles/using-pull-requests.
