+++ aem-component-generator/src/main/java/com/adobe/aem/compgenerator/Constants.java
@@ -1,0 +1,76 @@
+/*
+ * #%L
+ * AEM Component Generator
+ * %%
+ * Copyright (C) 2019 Adobe
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package com.adobe.aem.compgenerator;
+
+public class Constants {
+
+    public static String RESOURCE_TYPE_TEXTFIELD = "granite/ui/components/coral/foundation/form/textfield";
+    public static String RESOURCE_TYPE_NUMBER = "granite/ui/components/coral/foundation/form/numberfield";
+    public static String RESOURCE_TYPE_CHECKBOX = "granite/ui/components/coral/foundation/form/checkbox";
+    public static String RESOURCE_TYPE_PATHFIELD = "granite/ui/components/coral/foundation/form/pathfield";
+    public static String RESOURCE_TYPE_TEXTAREA = "granite/ui/components/coral/foundation/form/textarea";
+    public static String RESOURCE_TYPE_HIDDEN = "granite/ui/components/coral/foundation/form/hidden";
+    public static String RESOURCE_TYPE_DATEPICKER = "granite/ui/components/coral/foundation/form/datepicker";
+    public static String RESOURCE_TYPE_SELECT = "granite/ui/components/coral/foundation/form/select";
+    public static String RESOURCE_TYPE_RADIOGROUP = "granite/ui/components/coral/foundation/form/radiogroup";
+    public static String RESOURCE_TYPE_RADIO = "granite/ui/components/coral/foundation/form/radio";
+    public static String RESOURCE_TYPE_IMAGE = "cq/gui/components/authoring/dialog/fileupload";
+    public static String RESOURCE_TYPE_IMAGE_HIDDEN_TYPE = "core/wcm/components/image/v2/image";
+    public static String RESOURCE_TYPE_MULTIFIELD = "granite/ui/components/coral/foundation/form/multifield";
+    public static String RESOURCE_TYPE_DIALOG = "cq/gui/components/authoring/dialog";
+    public static String RESOURCE_TYPE_CONTAINER = "granite/ui/components/foundation/container";
+    public static String RESOURCE_TYPE_TABS = "granite/ui/components/foundation/layouts/tabs";
+    public static String RESOURCE_TYPE_FIXEDCOLUMNS = "granite/ui/components/foundation/layouts/fixedcolumns";
+    public static String RESOURCE_TYPE_SECTION = "granite/ui/components/foundation/section";
+    public static String RESOURCE_TYPE_FIELDSET = "granite/ui/components/foundation/form/fieldset";
+
+    public static String DIALOG_TYPE_DIALOG = "_cq_dialog";
+    public static String DIALOG_TYPE_GLOBAL = "dialogglobal";
+    public static String DIALOG_TYPE_SHARED = "dialogshared";
+    public static String FILENAME_CONTENT_XML = ".content.xml";
+
+    public static String PROPERTY_TYPE_PRIVATE = "private";
+    public static String PROPERTY_TYPE_GLOBAL = "global";
+    public static String PROPERTY_TYPE_SHARED = "shared";
+
+    public static String PROPERTY_JCR_TITLE = "jcr:title";
+    public static String PROPERTY_SLING_RESOURCETYPE = "sling:resourceType";
+    public static String PROPERTY_FIELDLABEL = "fieldLabel";
+    public static String PROPERTY_FIELDDESC = "fieldDescription";
+    public static String PROPERTY_NAME = "name";
+    public static String PROPERTY_COMPOSITE = "composite";
+    public static String PROPERTY_CQ_MSM_LOCKABLE = "cq-msm-lockable";
+
+    public static String JCR_ROOT_NODE = "jcr:root";
+    public static String JCR_PRIMARY_TYPE = "jcr:primaryType";
+    public static String NT_UNSTRUCTURED = "nt:unstructured";
+    public static String TYPE_COMPONENT = "cq:Component";
+    public static String TYPE_SLING_FOLDER = "sling:Folder";
+
+    public static String TEMPLATE_COPYRIGHT_JAVA = "template-copyright.txt";
+    public static String TEMPLATE_COPYRIGHT_XML = "template-copyright-xml.txt";
+    public static String TEMPLATE_HTL = "template-htl.txt";
+    public static final String TEMPLATE_COPYRIGHT_CSS = "template-copyright-css.txt";
+
+    public static String STRING_GET = "get";
+
+    public static final String RENAME_FILE_DATE_PATTERN = "yyyyMMddHHmmss";
+
+}
