+++ aem-component-generator/src/main/java/com/adobe/aem/compgenerator/exceptions/GeneratorException.java
@@ -1,0 +1,30 @@
+/*
+ * #%L
+ * AEM Component Generator
+ * %%
+ * Copyright (C) 2019 Adobe
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package com.adobe.aem.compgenerator.exceptions;
+
+public class GeneratorException extends RuntimeException {
+
+    private static final long serialVersionUID = -548600394249617384L;
+
+    public GeneratorException(String errorMessage) {
+        super(errorMessage);
+    }
+
+}
