--- aem-component-generator/src/main/java/com/bounteous/aem/compgenerator/exceptions/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Package contains the exceptions provided by the application.
- * @since 1.0
- */
-package com.bounteous.aem.compgenerator.exceptions;
