--- aem-component-generator/src/main/java/com/bounteous/aem/compgenerator/javacodemodel/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Package contains the java code generation classes.
- * @since 1.0
- */
-package com.bounteous.aem.compgenerator.javacodemodel;
