--- aem-component-generator/src/main/java/com/bounteous/aem/compgenerator/models/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Package contains model classes used by the application.
- * @since 1.0
- */
-package com.bounteous.aem.compgenerator.models;
