--- aem-component-generator/src/main/java/com/bounteous/aem/compgenerator/package-info.java
@@ -1,5 +1,0 @@
-/**
- * This is the core package for the application
- * @since 1.0
- */
-package com.bounteous.aem.compgenerator;
