+++ aem-component-generator/src/main/resources/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.bounteous.aem.compgenerator.AemCompGenerator
+Main-Class: com.adobe.aem.compgenerator.AemCompGenerator
